Conflicts:
	StortSpelprojekt/StortSpelprojekt/BaseEdit.cpp
	StortSpelprojekt/StortSpelprojekt/ObjectHandler.cpp
	StortSpelprojekt/StortSpelprojekt/StateMachine/LevelEditState.cpp

+++ StortSpelprojekt/StortSpelprojekt/Game.h
@@ -50,7 +50,7 @@ private:
 	//Resizing window, directx resources, camera
 	void ResizeResources(System::Settings* settings);
 
-	bool Update(double deltaTime);
+	bool Update(float deltaTime);
 	void Render();
 
 	void RenderGameObjects(int forShaderStage, std::vector<std::vector<GameObject*>>* gameObjects);
